Change: do not assume spender shares part of the sum
In other words, the transaction

    Ben 20 Reed

would charge 10 to Ben and 10 to Reed. This is no longer the case!
Now 20 is charged to Reed. This makes Fairsplit significantly more
useful.

Additionally, the command line interface has been a little bit more
polished, with a usage message and a few minor statements in other
places.

This is going to be release version 1.1.